requirements.txt
qdrant-client
openai-clip
streamlit
rawpy
exifread
transformers
torch
torchvision
pillow
accelerate
bitsandbytes 